Implement HTTP transport dockerized test
