+++ .circleci/config.yml
@@ -7,4 +7,7 @@ jobs:
       - checkout
       - run:
           name: Run AMQP integration specs
-          command: docker-compose up --abort-on-container-exit protein-client
+          command: docker-compose up --abort-on-container-exit protein-client-amqp
+      - run:
+          name: Run HTTP integration specs
+          command: docker-compose up --abort-on-container-exit protein-client-http
