+++ spec/integration/http_test_server.rb
@@ -1,0 +1,10 @@
+require_relative 'rpc_contract'
+
+require 'rack'
+require 'rack/show_exceptions'
+
+Rack::Server.start(
+  app: Protein::HTTPAdapter::Middleware.new(RPC::Router, ENV["RPC_SECRET"]),
+  Host: '0.0.0.0',
+  Port: '80'
+)
